Move video tile components in "tiles/video-tile" folder
