+++ libs/search-widget/src/tiles/video-tile/ParagraphPlayer.scss
