+++ libs/search-widget/src/tiles/video-tile/ThumbnailPlayer.scss
