+++ libs/search-widget/src/tiles/video-tile/TimePlayer.scss
