+++ libs/search-widget/src/tiles/video-tile/VideoTile.scss
